Crate musli_core

Core traits for Müsli.

Modules

buf

Types related to buffers.

de

Traits for generically dealing with a decoding framework.

en

Traits for generically dealing with an encoding framework.

hint

Core encoding hints.

mode

Type that describe a mode of encoding.

no_std

Trait fills for #[no_std] environments.

Traits

Allocator

An allocator that can be used in combination with a context.

Buf

A buffer allocated from a context.

Context

Provides ergonomic access to the serialization context.

Decode

Trait governing how types are decoded.

Decoder

Trait governing the implementation of a decoder.

Encode

Trait governing how types are encoded.

Encoder

Trait governing how the encoder works.

Attribute Macros

decoder

This is an attribute macro that must be used when implementing a Decoder.

encoder

This is an attribute macro that must be used when implementing a Encoder.

visitor

This is an attribute macro that must be used when implementing a Visitor.

Derive Macros

Decode

Derive which automatically implements the Decode trait.

Encode

Derive which automatically implements the Encode trait.