musli_core

Module alloc

Müsli Rust core allocation and collections library

This library provides smart pointers and collections for managing allocated values. This is similar to the Rust alloc crate, it provides similar but more limited functionality. However it can do so safely using the Müsli-specific Allocator trait allowing these types to be used in no_std environments without requiring a nightly compiler or unsafe.

Structs

AllocError

An allocation error.

Box

A Müsli-allocated pointer type that uniquely owns a heap allocation of type T.

Disabled

An allocator which cannot allocate anything.

FromUtf8Error

A possible error value when converting a String from a UTF-8 byte vector.

String

A Müsli-allocated UTF-8–encoded, growable string.

System

System buffer that can be used in combination with an Allocator.

SystemAlloc

A vector-backed allocation.

Vec

A Müsli-allocated contiguous growable array type, written as Vec<T>, short for ‘vector’.

Traits

Alloc

A value allocated through Allocator::alloc.

Allocator

An allocator that can be used in combination with a context.

ToOwned

The local `ToOwned`` implementation for Musli’s allocation system.