musli_core::buf
Trait Buf

pub trait Buf {
    // Required methods
    fn write(&mut self, bytes: &[u8]) -> bool;
    fn len(&self) -> usize;
    fn as_slice(&self) -> &[u8] ;
    fn write_fmt(&mut self, arguments: Arguments<'_>) -> Result<(), Error>;

    // Provided methods
    fn write_buffer<B>(&mut self, other: B) -> bool
       where B: Buf { ... }
    fn push(&mut self, byte: u8) -> bool { ... }
    fn is_empty(&self) -> bool { ... }
}

A buffer allocated from a context.

Buffers are allocated through an allocator using Allocator::alloc.

Required Methods

fn write(&mut self, bytes: &[u8]) -> bool

Write the given number of bytes.

Returns true if the bytes could be successfully written. A false value indicates that we are out of buffer capacity.

fn len(&self) -> usize

Get the length of the buffer in bytes.

fn as_slice(&self) -> &[u8]

Get the buffer as its initialized slice.

fn write_fmt(&mut self, arguments: Arguments<'_>) -> Result<(), Error>

Try to write a format string into the buffer.

Provided Methods

fn write_buffer<B>(&mut self, other: B) -> bool
where B: Buf,

Write a buffer of the same type onto the current buffer.

This allows allocators to provide more efficient means of extending the current buffer with one provided from the same allocator.

fn push(&mut self, byte: u8) -> bool

Write a single byte.

Returns true if the bytes could be successfully written. A false value indicates that we are out of buffer capacity.

fn is_empty(&self) -> bool

Test if the buffer is empty.

Object Safety

This trait is not object safe.