Trait musli_core::de::NumberVisitor

pub trait NumberVisitor<'de, C: ?Sized + Context>: Sized {
    type Ok;

    // Required method
    fn expecting(&self, f: &mut Formatter<'_>) -> Result;

    // Provided methods
    fn visit_u8(self, cx: &C, _: u8) -> Result<Self::Ok, C::Error> { ... }
    fn visit_u16(self, cx: &C, _: u16) -> Result<Self::Ok, C::Error> { ... }
    fn visit_u32(self, cx: &C, _: u32) -> Result<Self::Ok, C::Error> { ... }
    fn visit_u64(self, cx: &C, _: u64) -> Result<Self::Ok, C::Error> { ... }
    fn visit_u128(self, cx: &C, _: u128) -> Result<Self::Ok, C::Error> { ... }
    fn visit_i8(self, cx: &C, _: i8) -> Result<Self::Ok, C::Error> { ... }
    fn visit_i16(self, cx: &C, _: i16) -> Result<Self::Ok, C::Error> { ... }
    fn visit_i32(self, cx: &C, _: i32) -> Result<Self::Ok, C::Error> { ... }
    fn visit_i64(self, cx: &C, _: i64) -> Result<Self::Ok, C::Error> { ... }
    fn visit_i128(self, cx: &C, _: i128) -> Result<Self::Ok, C::Error> { ... }
    fn visit_f32(self, cx: &C, _: f32) -> Result<Self::Ok, C::Error> { ... }
    fn visit_f64(self, cx: &C, _: f64) -> Result<Self::Ok, C::Error> { ... }
    fn visit_usize(self, cx: &C, _: usize) -> Result<Self::Ok, C::Error> { ... }
    fn visit_isize(self, cx: &C, _: isize) -> Result<Self::Ok, C::Error> { ... }
    fn visit_bytes(self, cx: &C, _: &'de [u8]) -> Result<Self::Ok, C::Error> { ... }

}

A visitor capable of processing arbitrary number values.

Required Associated Types

type Ok

The output of the visitor.

Required Methods

fn expecting(&self, f: &mut Formatter<'_>) -> Result

Format an error indicating what was expected by this visitor.

Override to be more specific about the type that failed.

Provided Methods

fn visit_u8(self, cx: &C, _: u8) -> Result<Self::Ok, C::Error>

Visit u8.

fn visit_u16(self, cx: &C, _: u16) -> Result<Self::Ok, C::Error>

Visit u16.

fn visit_u32(self, cx: &C, _: u32) -> Result<Self::Ok, C::Error>

Visit u32.

fn visit_u64(self, cx: &C, _: u64) -> Result<Self::Ok, C::Error>

Visit u64.

fn visit_u128(self, cx: &C, _: u128) -> Result<Self::Ok, C::Error>

Visit u128.

fn visit_i8(self, cx: &C, _: i8) -> Result<Self::Ok, C::Error>

Visit i8.

fn visit_i16(self, cx: &C, _: i16) -> Result<Self::Ok, C::Error>

Visit i16.

fn visit_i32(self, cx: &C, _: i32) -> Result<Self::Ok, C::Error>

Visit i32.

fn visit_i64(self, cx: &C, _: i64) -> Result<Self::Ok, C::Error>

Visit i64.

fn visit_i128(self, cx: &C, _: i128) -> Result<Self::Ok, C::Error>

Visit i128.

fn visit_f32(self, cx: &C, _: f32) -> Result<Self::Ok, C::Error>

Visit f32.

fn visit_f64(self, cx: &C, _: f64) -> Result<Self::Ok, C::Error>

Visit f64.

fn visit_usize(self, cx: &C, _: usize) -> Result<Self::Ok, C::Error>

Visit usize.

fn visit_isize(self, cx: &C, _: isize) -> Result<Self::Ok, C::Error>

Visit isize.

fn visit_bytes(self, cx: &C, _: &'de [u8]) -> Result<Self::Ok, C::Error>

Visit bytes constituting a raw number.

Object Safety

This trait is not object safe.