musli_core::de

Trait DecodeTrace

pub trait DecodeTrace<'de, M, A>: Sized
where
    A: Allocator,
{
    // Required method
    fn trace_decode<D>(decoder: D) -> Result<Self, D::Error>
       where D: Decoder<'de, Mode = M, Allocator = A>;
}

Trait governing how types are decoded specifically for tracing.

This is used for types where some extra bounds might be necessary to trace a container such as a HashMap<K, V> where K would have to implement fmt::Display.

Required Methods

fn trace_decode<D>(decoder: D) -> Result<Self, D::Error>
where D: Decoder<'de, Mode = M, Allocator = A>,

Decode the given input.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.