musli_core::de

Trait SequenceDecoder

pub trait SequenceDecoder<'de> {
    type Cx: Context<Error = Self::Error, Allocator = Self::Allocator>;
    type Error;
    type Allocator: Allocator;
    type Mode: 'static;
    type DecodeNext<'this>: Decoder<'de, Cx = Self::Cx, Error = Self::Error, Allocator = Self::Allocator, Mode = Self::Mode>
       where Self: 'this;

    // Required methods
    fn cx(&self) -> Self::Cx;
    fn decode_next(&mut self) -> Result<Self::DecodeNext<'_>, Self::Error>;
    fn try_decode_next(
        &mut self,
    ) -> Result<Option<Self::DecodeNext<'_>>, Self::Error>;

    // Provided methods
    fn size_hint(&self) -> SizeHint { ... }
    fn next<T>(&mut self) -> Result<T, Self::Error>
       where T: Decode<'de, Self::Mode, Self::Allocator> { ... }
    fn try_next<T>(&mut self) -> Result<Option<T>, Self::Error>
       where T: Decode<'de, Self::Mode, Self::Allocator> { ... }
}

Trait governing how to decode a sequence.

Required Associated Types

type Cx: Context<Error = Self::Error, Allocator = Self::Allocator>

Context associated with the decoder.

type Error

Error associated with decoding.

type Allocator: Allocator

The allocator associated with the decoder.

type Mode: 'static

The mode of the decoder.

type DecodeNext<'this>: Decoder<'de, Cx = Self::Cx, Error = Self::Error, Allocator = Self::Allocator, Mode = Self::Mode> where Self: 'this

The decoder for individual items.

Required Methods

fn cx(&self) -> Self::Cx

Access the context associated with the decoder.

fn decode_next(&mut self) -> Result<Self::DecodeNext<'_>, Self::Error>

Return decoder to decode the next element.

This will error or provide garbled data in case the next element is not available.

fn try_decode_next( &mut self, ) -> Result<Option<Self::DecodeNext<'_>>, Self::Error>

Try to decode the next element.

Provided Methods

fn size_hint(&self) -> SizeHint

Get a size hint of known remaining elements.

fn next<T>(&mut self) -> Result<T, Self::Error>
where T: Decode<'de, Self::Mode, Self::Allocator>,

Decode the next element of the given type, erroring in case it’s absent.

fn try_next<T>(&mut self) -> Result<Option<T>, Self::Error>
where T: Decode<'de, Self::Mode, Self::Allocator>,

Decode the next element of the given type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.